suppression de is_blocked_bonus.c , amelioration parsing porte, simplification fonctions move
