+++ README.md
@@ -47,9 +47,11 @@ make bonus
 ```bash
 # Version standard
 ./cub3D maps/map1.cub
+./cub3D maps/map2.cub
+./cub3D maps/map3.cub
 
 # Version avec bonus
-./cub3D_bonus maps/map4.cub
+./cub3D_bonus maps/map_bonus.cub
 ```
 
 ### Format du fichier de carte (.cub)
