+++ src/move.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   move.c                                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: cparodi <cparodi@student.42.fr>            +#+  +:+       +#+        */
+/*   By: ayarmaya <ayarmaya@student.42nice.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/27 16:18:00 by cparodi           #+#    #+#             */
-/*   Updated: 2025/04/27 16:18:02 by cparodi          ###   ########.fr       */
+/*   Updated: 2025/05/03 23:06:46 by ayarmaya         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,80 +16,52 @@ void	move_forward(t_game *game)
 {
 	double	new_x;
 	double	new_y;
-	int		map_x;
-	int		map_y;
 
 	new_x = game->player.pos.x + game->player.dir.x * MOVE_SPEED;
 	new_y = game->player.pos.y + game->player.dir.y * MOVE_SPEED;
-	map_x = (int)new_x;
-	map_y = (int)new_y;
-	if (game->map[map_y][map_x] == '1')
-	{
-		if (is_exterior_wall(game, map_x, map_y))
-			return ;
-	}
-	game->player.pos.x = new_x;
-	game->player.pos.y = new_y;
+	if (!is_exterior_wall(game, (int)new_x, (int)game->player.pos.y))
+		game->player.pos.x = new_x;
+	if (!is_exterior_wall(game, (int)game->player.pos.x, (int)new_y))
+		game->player.pos.y = new_y;
 }
 
 void	move_backward(t_game *game)
 {
 	double	new_x;
 	double	new_y;
-	int		map_x;
-	int		map_y;
 
 	new_x = game->player.pos.x - game->player.dir.x * MOVE_SPEED;
 	new_y = game->player.pos.y - game->player.dir.y * MOVE_SPEED;
-	map_x = (int)new_x;
-	map_y = (int)new_y;
-	if (game->map[map_y][map_x] == '1')
-	{
-		if (is_exterior_wall(game, map_x, map_y))
-			return ;
-	}
-	game->player.pos.x = new_x;
-	game->player.pos.y = new_y;
+	if (!is_exterior_wall(game, (int)new_x, (int)game->player.pos.y))
+		game->player.pos.x = new_x;
+	if (!is_exterior_wall(game, (int)game->player.pos.x, (int)new_y))
+		game->player.pos.y = new_y;
 }
 
 void	move_left(t_game *game)
 {
 	double	new_x;
 	double	new_y;
-	int		map_x;
-	int		map_y;
 
 	new_x = game->player.pos.x + game->player.dir.y * MOVE_SPEED;
 	new_y = game->player.pos.y - game->player.dir.x * MOVE_SPEED;
-	map_x = (int)new_x;
-	map_y = (int)new_y;
-	if (game->map[map_y][map_x] == '1')
-	{
-		if (is_exterior_wall(game, map_x, map_y))
-			return ;
-	}
-	game->player.pos.x = new_x;
-	game->player.pos.y = new_y;
+	if (!is_exterior_wall(game, (int)new_x, (int)game->player.pos.y))
+		game->player.pos.x = new_x;
+	if (!is_exterior_wall(game, (int)game->player.pos.x, (int)new_y))
+		game->player.pos.y = new_y;
 }
 
 void	move_right(t_game *game)
 {
 	double	new_x;
 	double	new_y;
-	int		map_x;
-	int		map_y;
 
 	new_x = game->player.pos.x - game->player.dir.y * MOVE_SPEED;
 	new_y = game->player.pos.y + game->player.dir.x * MOVE_SPEED;
-	map_x = (int)new_x;
-	map_y = (int)new_y;
-	if (game->map[map_y][map_x] == '1')
-	{
-		if (is_exterior_wall(game, map_x, map_y))
-			return ;
-	}
-	game->player.pos.x = new_x;
-	game->player.pos.y = new_y;
+	if (!is_exterior_wall(game, (int)new_x, (int)game->player.pos.y))
+		game->player.pos.x = new_x;
+	if (!is_exterior_wall(game, (int)game->player.pos.x, (int)new_y))
+		game->player.pos.y = new_y;
 }
 
 void	rotate(t_game *game, double angle)
